Reworked repo to use NPM/browserify, removed Bower
