--- wp-content/themes/vtg-theme/Gemfile
@@ -1,5 +1,0 @@
-source "https://rubygems.org"
-
-gem 'susy', '~> 2.2.5'
-gem 'compass', '~> 1.0.3'
-gem 'normalize-scss', '3.0.2'
